Refactor link selection in aggtreecluster
Alter the way the link variable is processed so that we first explicitly check for its type (str, float, ndarray, or function).  This way when we start subdividing the possibilities within str, we don't end up trying to compare across types (which can introduce errors, particularly for ndarray where element-by-element checking is possible).
Also, for some reason the centroid based methods were under type(link) is byte heading, when the options used to select those methods are actually strings.  This has been fixed so that they appear in the string section.