Test result differentiation
Test failures are differentiated into three types:
 - Failures that result from the function being tested raising an exception
 - Failures of non-numerical tests which are conducted on a pure pass/fail basis
 - Failures of numerical tests in which results are outside of the allowable tolerance (defined by rtol and atol)

Formerly tests which raised exceptions simply crashed the testing algorithm, and this is still the default behavior.  However a `force` argument has been added to the testing functions which will allow the testing to continue when an exception is encountered so that a more complete picture of the test results can be gathered before starting to fix the problems identified by the tests.

I've also discovered that that when a low-level function (such as distance) which is called by higher-level functions (such as aggtreecluster) failed due to an exception, this can result in the higher-level function failing without raising an exception.  This will have to be investigated and fixed.