Tests for treebuild
We use the imagehash and Pillow (a fork of PIL, not the original PIL) libraries to compare the test image to the standard image.
Higher values of image_hash_size allow for a more detailed comparison.  At the default (8) the difference between the image standards and the test images on the same computer is 0 (as expected).
The default value for image_hash_diff is a guess as to how much differences should be allowed.  I will need to run the tests on different computers in order to get a sense of just how much difference should actually be allowed.